feat(seeders): create userSeeder, snippetSeeder, seed and clearDatabase files
