+++ src/controllers/userController.ts
@@ -5,90 +5,6 @@ import { UserType } from "../models/User";
 
 const userRepository = AppDataSource.getRepository(User);
 
-const users: UserType[] = [
-  {
-    id: 1,
-    firstname: "John",
-    lastname: "Doe",
-    email: "john@example.com",
-    pseudo: "johndoe",
-    password: "password123",
-  },
-  {
-    id: 2,
-    firstname: "Jane",
-    lastname: "Doe",
-    email: "jane@example.com",
-    pseudo: "janedoe",
-    password: "password456",
-  },
-
-  {
-    id: 3,
-    firstname: "Alice",
-    lastname: "Smith",
-    email: "alice@example.com",
-    pseudo: "alicesmith",
-    password: "password789",
-  },
-  {
-    id: 4,
-    firstname: "Bob",
-    lastname: "Johnson",
-    email: "bob@example.com",
-    pseudo: "bobjohnson",
-    password: "password101",
-  },
-  {
-    id: 5,
-    firstname: "Charlie",
-    lastname: "Brown",
-    email: "charlie@example.com",
-    pseudo: "charliebrown",
-    password: "password102",
-  },
-  {
-    id: 6,
-    firstname: "David",
-    lastname: "Miller",
-    email: "david@example.com",
-    pseudo: "davidmiller",
-    password: "password103",
-  },
-  {
-    id: 7,
-    firstname: "Emma",
-    lastname: "Wilson",
-    email: "emma@example.com",
-    pseudo: "emmawilson",
-    password: "password104",
-  },
-  {
-    id: 8,
-    firstname: "Frank",
-    lastname: "Taylor",
-    email: "frank@example.com",
-    pseudo: "franktaylor",
-    password: "password105",
-  },
-  {
-    id: 9,
-    firstname: "Grace",
-    lastname: "Anderson",
-    email: "grace@example.com",
-    pseudo: "graceanderson",
-    password: "password106",
-  },
-  {
-    id: 10,
-    firstname: "Henry",
-    lastname: "Thomas",
-    email: "henry@example.com",
-    pseudo: "henrythomas",
-    password: "password107",
-  },
-];
-
 export const getUsers = async (req: Request, res: Response<UserType[]>) => {
   const users = await userRepository.find();
